harambee-backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-cors==4.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.21.3
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
